+++ src/rpc/client.rs
@@ -3,29 +3,29 @@
 
 //! # Design Goals
 //! - use [`jsonrpsee`] clients and primitives.
-//! - Support different call formats
-//!   - [`crate::rpc_client::RpcRequest`]
-//!   - [`crate::rpc::RpcMethod`]
+//! - Support [`rpc::Request`](crate::rpc::Request).
 //! - Support different
 //!   - endpoint paths (`v0`, `v1`).
 //!   - communication protocols (`ws`, `http`).
 //! - Support per-request timeouts.
 
+use std::env;
 use std::fmt::{self, Debug};
 use std::time::Duration;
 
+use anyhow::bail;
 use http02::{header, HeaderMap, HeaderValue};
 use jsonrpsee::core::client::ClientT as _;
 use jsonrpsee::core::params::{ArrayParams, ObjectParams};
 use jsonrpsee::core::ClientError;
+use once_cell::sync::Lazy;
 use serde::de::DeserializeOwned;
 use tracing::{debug, Instrument, Level};
 use url::Url;
 
-use super::{ApiVersion, MAX_REQUEST_BODY_SIZE, MAX_RESPONSE_BODY_SIZE};
+use super::{ApiVersion, Request, MAX_REQUEST_BODY_SIZE, MAX_RESPONSE_BODY_SIZE};
 
-/// A JSON-RPC client that can dispatch either a [`crate::rpc_client::RpcRequest`]
-/// or a [`crate::rpc::RpcMethod`] to a single URL.
+/// A JSON-RPC client that can dispatch either a [`crate::rpc::Request`] to a single URL.
 pub struct Client {
     /// SHOULD end in a slash, due to our use of [`Url::join`].
     base_url: Url,
@@ -36,19 +36,43 @@ pub struct Client {
 }
 
 impl Client {
-    pub fn new(base_url: Url, token: impl Into<Option<String>>) -> Self {
+    /// Use either the URL in the environment or a default.
+    ///
+    /// If `token` is provided, use that over the token in either of the above.
+    pub fn default_or_from_env(token: Option<&str>) -> anyhow::Result<Self> {
+        static DEFAULT: Lazy<Url> = Lazy::new(|| "http://127.0.0.1:2345/".parse().unwrap());
+
+        let mut base_url = match env::var("FULLNODE_API_INFO") {
+            Ok(it) => {
+                let crate::utils::UrlFromMultiAddr(url) = it.parse()?;
+                url
+            }
+            Err(env::VarError::NotPresent) => DEFAULT.clone(),
+            Err(e @ env::VarError::NotUnicode(_)) => bail!(e),
+        };
+        if token.is_some() && base_url.set_password(token).is_err() {
+            bail!("couldn't set override password")
+        }
+        Ok(Self::from_url(base_url))
+    }
+    pub fn from_url(mut base_url: Url) -> Self {
+        let token = base_url.password().map(Into::into);
+        let _defer = base_url.set_password(None);
         Self {
+            token,
             base_url,
-            token: token.into(),
             v0: Default::default(),
             v1: Default::default(),
         }
     }
+    pub fn base_url(&self) -> &Url {
+        &self.base_url
+    }
     pub async fn call<T: crate::lotus_json::HasLotusJson + std::fmt::Debug>(
         &self,
-        req: crate::rpc_client::RpcRequest<T>,
+        req: Request<T>,
     ) -> Result<T, ClientError> {
-        let crate::rpc_client::RpcRequest {
+        let Request {
             method_name,
             params,
             api_version,
@@ -140,7 +164,7 @@ fn trace_params(params: impl jsonrpsee::core::traits::ToRpcParams) {
 /// can be made using [`jsonrpsee`] primitives.
 struct UrlClient {
     url: Url,
-    inner: OneClientInner,
+    inner: UrlClientInner,
 }
 
 impl Debug for UrlClient {
@@ -169,7 +193,7 @@ impl UrlClient {
             None => HeaderMap::new(),
         };
         let inner = match url.scheme() {
-            "ws" | "wss" => OneClientInner::Ws(
+            "ws" | "wss" => UrlClientInner::Ws(
                 jsonrpsee::ws_client::WsClientBuilder::new()
                     .set_headers(headers)
                     .max_request_size(MAX_REQUEST_BODY_SIZE)
@@ -178,7 +202,7 @@ impl UrlClient {
                     .build(&url)
                     .await?,
             ),
-            "http" | "https" => OneClientInner::Https(
+            "http" | "https" => UrlClientInner::Https(
                 jsonrpsee::http_client::HttpClientBuilder::new()
                     .set_headers(headers)
                     .max_request_size(MAX_REQUEST_BODY_SIZE)
@@ -197,7 +221,7 @@ impl UrlClient {
     }
 }
 
-enum OneClientInner {
+enum UrlClientInner {
     Ws(jsonrpsee::ws_client::WsClient),
     Https(jsonrpsee::http_client::HttpClient),
 }
@@ -210,8 +234,8 @@ impl jsonrpsee::core::client::ClientT for UrlClient {
         params: P,
     ) -> Result<(), jsonrpsee::core::ClientError> {
         match &self.inner {
-            OneClientInner::Ws(it) => it.notification(method, params).await,
-            OneClientInner::Https(it) => it.notification(method, params).await,
+            UrlClientInner::Ws(it) => it.notification(method, params).await,
+            UrlClientInner::Https(it) => it.notification(method, params).await,
         }
     }
     async fn request<R: DeserializeOwned, P: jsonrpsee::core::traits::ToRpcParams + Send>(
@@ -220,8 +244,8 @@ impl jsonrpsee::core::client::ClientT for UrlClient {
         params: P,
     ) -> Result<R, jsonrpsee::core::ClientError> {
         match &self.inner {
-            OneClientInner::Ws(it) => it.request(method, params).await,
-            OneClientInner::Https(it) => it.request(method, params).await,
+            UrlClientInner::Ws(it) => it.request(method, params).await,
+            UrlClientInner::Https(it) => it.request(method, params).await,
         }
     }
     async fn batch_request<'a, R: DeserializeOwned + 'a + std::fmt::Debug>(
@@ -229,8 +253,8 @@ impl jsonrpsee::core::client::ClientT for UrlClient {
         batch: jsonrpsee::core::params::BatchRequestBuilder<'a>,
     ) -> Result<jsonrpsee::core::client::BatchResponse<'a, R>, jsonrpsee::core::ClientError> {
         match &self.inner {
-            OneClientInner::Ws(it) => it.batch_request(batch).await,
-            OneClientInner::Https(it) => it.batch_request(batch).await,
+            UrlClientInner::Ws(it) => it.batch_request(batch).await,
+            UrlClientInner::Https(it) => it.batch_request(batch).await,
         }
     }
 }
@@ -248,11 +272,11 @@ impl jsonrpsee::core::client::SubscriptionClientT for UrlClient {
         Notif: DeserializeOwned,
     {
         match &self.inner {
-            OneClientInner::Ws(it) => {
+            UrlClientInner::Ws(it) => {
                 it.subscribe(subscribe_method, params, unsubscribe_method)
                     .await
             }
-            OneClientInner::Https(it) => {
+            UrlClientInner::Https(it) => {
                 it.subscribe(subscribe_method, params, unsubscribe_method)
                     .await
             }
@@ -266,8 +290,8 @@ impl jsonrpsee::core::client::SubscriptionClientT for UrlClient {
         Notif: DeserializeOwned,
     {
         match &self.inner {
-            OneClientInner::Ws(it) => it.subscribe_to_method(method).await,
-            OneClientInner::Https(it) => it.subscribe_to_method(method).await,
+            UrlClientInner::Ws(it) => it.subscribe_to_method(method).await,
+            UrlClientInner::Https(it) => it.subscribe_to_method(method).await,
         }
     }
 }