+++ src/rpc/types/mod.rs
@@ -41,8 +41,8 @@ use fil_actors_shared::fvm_ipld_bitfield::BitField;
 use fvm_ipld_encoding::{BytesDe, RawBytes};
 use libipld_core::ipld::Ipld;
 use libp2p::PeerId;
-use nonempty::NonEmpty;
 use num_bigint::BigInt;
+use nunny::Vec as NonEmpty;
 use schemars::JsonSchema;
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 #[cfg(test)]
@@ -327,7 +327,7 @@ pub struct SectorOnChainInfo {
 
 lotus_json_with_self!(SectorOnChainInfo);
 
-#[derive(Clone, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "PascalCase")]
 pub struct SectorPreCommitOnChainInfo {
     #[schemars(with = "LotusJson<SectorPreCommitInfo>")]
@@ -341,7 +341,7 @@ pub struct SectorPreCommitOnChainInfo {
 
 lotus_json_with_self!(SectorPreCommitOnChainInfo);
 
-#[derive(Clone, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "PascalCase")]
 pub struct SectorPreCommitInfo {
     #[schemars(with = "LotusJson<RegisteredSealProof>")]